backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pyyaml>=6.0
aiohttp>=3.9.0
cryptography>=42.0
python-dotenv>=1.0
watchfiles>=1.0
